Refactor common code into a globals file.  Update some text to be more user friendly and refactor main page to be breed trends.
