+++ Procfile
@@ -1,1 +1,1 @@
-web: sh setup.sh && streamlit run projects/rescuechi/petfinder-streamlit/pet-analytics.py
+web: sh setup.sh && streamlit run projects/rescuechi/petfinder-streamlit/Breed_Trends.py
